+++ NEWS.md
@@ -1,5 +1,10 @@
 # cbcTools (development version)
 
+# cbcTools 0.3.0
+
+- Bug fix: modify code in `cbc_design()` to avoid duplicate choice sets for the same respondents; addresses #7.
+- Bug fix: modify code in `cbc_design()` to allow Bayesian D-efficient designs with restricted profile sets; addresses #10 and #9.
+
 # cbcTools 0.2.0
 
 - Added a startup message when the package is loaded.